Prepare tests to be reused across implementations
