+++ minireq/package.json
@@ -7,12 +7,12 @@
   "typings": "build/index.d.ts",
   "scripts": {
     "build": "tsc --project tsconfig.build.json && pnpm run build:esm && pnpm run build:cjs",
-    "build:esm": "rollup build/index.js --format esm --file build/bundle.esm.js",
-    "build:cjs": "rollup build/index.js --format cjs --file build/bundle.cjs.js",
+    "build:esm": "BUNDLE_FORMAT=esm rollup -c ../configs/rollup.config.ts",
+    "build:cjs": "BUNDLE_FORMAT=cjs rollup -c ../configs/rollup.config.ts",
     "test": "start-server-and-test start-server 3000 test:normal",
     "debug": "DEBUG=1 pnpm test",
     "test:normal": "karma start",
-    "start-server": "ts-node test/support/server.ts",
+    "start-server": "ts-node ../test/support/server.ts",
     "docs": "rm -r docs; typedoc --out docs --mode file --excludeNotExported src && touch docs/.nojekyll",
     "prepublishOnly": "pnpm run build && pnpm test",
     "format": "prettier --write './{src,test}/**/*.ts'"
@@ -22,29 +22,29 @@
   "devDependencies": {
     "@types/assert": "^1.4.6",
     "@types/cors": "^2.8.6",
-    "@types/express": "^4.17.2",
-    "@types/mocha": "^7.0.1",
-    "@types/node": "^13.7.0",
+    "@types/express": "^4.17.6",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.13.2",
     "assert": "^2.0.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "karma": "^4.4.1",
+    "karma": "^5.0.2",
     "karma-chrome-launcher": "^3.1.0",
     "karma-cli": "^2.0.0",
     "karma-firefox-launcher": "^1.3.0",
-    "karma-mocha": "^1.3.0",
-    "karma-typescript": "^4.1.1",
-    "mocha": "^7.0.1",
-    "nyc": "^15.0.0",
-    "prettier": "^1.19.1",
-    "rollup": "^1.31.0",
-    "start-server-and-test": "^1.10.8",
-    "ts-node": "^8.6.2",
-    "typedoc": "^0.16.9",
-    "typescript": "^3.7.5"
+    "karma-mocha": "^2.0.0",
+    "karma-typescript": "^5.0.2",
+    "mocha": "^7.1.1",
+    "nyc": "^15.0.1",
+    "prettier": "^2.0.5",
+    "rollup": "^2.7.2",
+    "start-server-and-test": "^1.11.0",
+    "ts-node": "^8.9.0",
+    "typedoc": "^0.17.4",
+    "typescript": "^3.8.3"
   },
   "dependencies": {
-    "minireq-types": "workspace:^1.0.0"
+    "minireq-common": "workspace:^1.0.0"
   },
-  "prettier": "../.prettierrc.json"
+  "prettier": "../configs/.prettierrc.json"
 }
