+++ minireq/test/common.ts
@@ -1,0 +1,4 @@
+import { makeTestSuite } from '../../test/index';
+import { makeRequest } from '../src/index';
+
+makeTestSuite(makeRequest);
