+++ minireq/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../configs/tsconfig.json",
   "include": ["./src/**/*", "./test/**/*"]
 }
